Separate out LogRequestItem
