+++ edge-agent/src/Microsoft.Azure.Devices.Edge.Agent.IoTHub/stream/LogsStreamRequest.cs
@@ -39,16 +39,4 @@ namespace Microsoft.Azure.Devices.Edge.Agent.IoTHub.Stream
         [DefaultValue(LogsContentType.Text)]
         public LogsContentType ContentType { get; }
     }
-
-    public class LogRequestItem
-    {
-        public LogRequestItem(string id, ModuleLogFilter filter)
-        {
-            this.Id = Preconditions.CheckNonWhiteSpace(id, nameof(id));
-            this.Filter = filter ?? ModuleLogFilter.Empty;
-        }
-
-        [JsonProperty("id")]
-        public string Id { get; }
-    }
 }
